fix: never replace undefined with an empty state
Previously, the 'state' passed to the mergeReducer would be '{}' when
state === undefined, unless the action was @ngrx/store/init.

This breaks State interfaces because it prevents the reducer function
returned by createReducer from defaulting the state parameter to
initialState, which would be of the correct type.

If the localStorageSync meta reducer is applied in the root module,
this issue is hidden by INIT being the first action.

Given that the inner reducer is called unconditionally after the
mergeReducer and there is little reason to apply @ngrx/store/init
twice, we can use a dedicated action type named after this package to
get the value of initialState.